// src/shared/api/models/getApiDoctorsPortalDoctorsConnectionStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiDoctorsPortalDoctorsConnectionStatus =
    (typeof GetApiDoctorsPortalDoctorsConnectionStatus)[keyof typeof GetApiDoctorsPortalDoctorsConnectionStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiDoctorsPortalDoctorsConnectionStatus = {
    none: 'none',
    invited: 'invited',
    linked: 'linked',
    removed: 'removed',
    self_unlinked: 'self_unlinked',
    rejected: 'rejected',
} as const;